# src/types/response/Admin.response.ts
export interface SaveUpdateResAdmin {
    _id: string,
    FirstName: string,
    LastName: string,
    Designation: string,
    Cell: string,
    JoinDate: string,
    Address: string,
    createdAt?: string,
    updatedAt?: string
}